mega script/requirements.txt
pyyaml
terminaltables
rich
distro
flatdict
dotmap
requests
beautifulsoup4